Refactor KG agent integration and update dependencies
- Replace KGMaintainerAgent with KnowledgeAgent in finalize_agent.py
- Remove deprecated kg_maintainer_agent.py module
- Update imports and initialization to use new KnowledgeAgent
- Adjust test files to reflect agent refactoring
