+++ .gitignore
@@ -25,3 +25,4 @@ story-text.txt
 user_story_elements.ba*
 chorus.md
 story-test.txt
+.aider*
